moved bit-fiddling code to separate namespace bit changed most std::size_t instances to int

+++ src/bit/Bit.h
@@ -1,5 +1,4 @@
 #pragma once
-#include <cstddef>
 
 namespace dctl {
 namespace bit {
@@ -17,19 +16,19 @@ template<typename T> bool is_exclusive(T, T);           // T1 and T2 mutually ex
 // unary views
 template<typename T> T get_first(T);		        // least significant 1-bit
 template<typename T> T except_first(T);	                // most significant 1-bits
-template<typename T> size_t find_first(T);              // index of the least significant 1-bit
+template<typename T> int find_first(T);                 // index of the least significant 1-bit
 
-template<typename T> size_t index(T);		        // index of a set 1-bit
-template<typename T> size_t index_DeBruijn(T);          // index of a set 1-bit
-template<typename T> size_t index_lookup(T);            // index of a set 1-bit
+template<typename T> int index(T);		        // index of a set 1-bit
+template<typename T> int index_DeBruijn(T);             // index of a set 1-bit
+template<typename T> int index_lookup(T);               // index of a set 1-bit
 
-template<typename T> size_t count(T);		        // number of set 1-bits
-template<typename T> size_t count_Kernighan(T);         // number of set 1-bits
-template<typename T> size_t count_lookup(T);            // number of set 1-bits
+template<typename T> int count(T);		        // number of set 1-bits
+template<typename T> int count_Kernighan(T);            // number of set 1-bits
+template<typename T> int count_lookup(T);               // number of set 1-bits
 
 // binary views
-template<bool, typename T> T flood_fill(T, T, size_t);  // direction-wise flood-fill T1 over T2
-template<bool, typename T> T fill_loop(T, T, size_t);   // direction-wise flood-fill T1 over T2
+template<bool, typename T> T flood_fill(T, T, int);     // direction-wise flood-fill T1 over T2
+template<bool, typename T> T fill_loop(T, T, int);      // direction-wise flood-fill T1 over T2
 
 // unary modifiers
 template<typename T> void clear_first(T&);		// clear the least significant 1-bit
