+++ src/dxp/BackRequest.h
@@ -1,7 +1,6 @@
 #pragma once
-#include <cstddef>
 #include <memory>       // std::unique_ptr
-#include <string>
+#include <string>       // std::string
 #include "MessageInterface.h"
 
 namespace dctl {
@@ -11,9 +10,9 @@ class BackRequest: public MessageInterface
 {
 public:
         // views
-        size_t move_number() const;
+        int move_number() const;
         char side_to_move() const;
-        static std::string str(size_t, char);
+        static std::string str(int, char);
 
 private:
         // private constructor to enforce factory creation
@@ -22,14 +21,14 @@ private:
         // implementation
         virtual std::string header() const;
         virtual std::string body() const;
-        static std::string body(size_t, char);
+        static std::string body(int, char);
 
         static std::unique_ptr<MessageInterface> create(const std::string&);
         static const std::string HEADER;
         static const bool REGISTERED;
 
         // representation
-        size_t move_number_;
+        int move_number_;
         char side_to_move_;
 };
